re-enabled refactored IVP handler

+++ application/views/ivpform.php
@@ -192,7 +192,11 @@ echo form_textarea($f_misc['field']);
 ?>
 
 <?php
-echo '<p>' . form_reset('reset', 'Ångra') . form_submit('submit', 'Spara') . '</p>';
+if (isset($READONLY) && $READONLY == 'READONLY') {
+	echo '<p></p>';
+} else {
+	echo '<p>' . form_reset('reset', 'Ångra') . form_submit('submit', 'Spara') . '</p>';
+}
 echo form_close();
 ?>
 
@@ -202,7 +206,7 @@ echo form_close();
 <ul>
 <?php
 foreach ($editors as $ed) {
-	echo '<li>' . $ed['stamp'] . ' (<b>' . $ed['username'] . '</b>)</li>' . "\n";
+	echo '<li>' . $ed['stamp'] . ' (<b>' . $ed['username'] . '</b>) ' . $ed['event_type'] . '</li>' . "\n";
 }}
 ?>
 </ul></div>
